Added Doxygen comments and stuff for generating a nice reference.
Not that it is necessary, but it gives me good practice for what I'm doing with capstone documentation, plus it looks cool. I also did a little tidying up of the directory and fixed some issues with the build and clean files I made. I want to add generating doxy output into the build files but I haven't gotten that to work on command line yet on my machine. This commit doesn't really contribute any code to the project, its just posterity and organizational.