Client and Server Proxy
Added skeleton code to the client and server proxy classes. Also some changes to the class design/diagram is not up to date on this commit. IDK if anyone else made changes to update it. Alterations to this diagram will be made over the next day or two as we continue to design and develop the code and find things we need to change.