Client Now Threads And Calls Transactions
Some fixes to the command line arguments parsing (will likely do more and even think about reading a config file instead). Client now generates random numbers for account numbers and amount to transfer between accounts. Worker threads are implemented to run the transactions based on these numbers. Appear to correctly call proxy, which appears to correctly communicate with server, which appears to correctly respond (with dummy data and without actually performing the steps of the transactions). More changes will come to the client side to refine things, but for now consider this working.